Added a 'smart' genetic generator
The generator favors its random functions to select libraries
with a good potential. Since the calculation of the potential
of libraries is the same as the search generator this ends up
being just a random function that tends to the same results as
the search generator. Still an interesting PoC anyway.